@@ -19,19 +19,19 @@
   "bsb_project": true,
   "devDependencies": {
     "@types/crypto-js": "^4.1.1",
-    "@typescript-eslint/eslint-plugin": "^5.15.0",
-    "@typescript-eslint/parser": "^5.15.0",
-    "eslint": "^8.11.0",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.2"
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
+    "eslint": "^8.19.0",
+    "ts-node": "^10.8.2",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
-    "@bettercorp/service-base": "^7.6.20220316220708",
-    "@bettercorp/service-base-plugin-web-server": "^4.1.20220316224011",
-    "@bettercorp/tools": "^1.3.20220310075141",
-    "@types/node": "^17.0.21",
+    "@bettercorp/service-base": "^7.6.20220531162851",
+    "@bettercorp/service-base-plugin-web-server": "^4.1.20220429033429",
+    "@bettercorp/tools": "^2.0.20220629134925",
+    "@types/node": "^18.0.3",
     "@types/uuid": "^8.3.4",
-    "axios": "^0.26.1",
-    "moment": "^2.29.1"
+    "axios": "^0.27.2",
+    "moment": "^2.29.4"
   }
 }
