@@ -1,12 +1,12 @@
-import { Tools } from '@bettercorp/tools/lib/Tools';
-import { IServerConfig, webRequest as CWR } from '../../weblib';
-import moment = require('moment');
-import { IDictionary } from '@bettercorp/tools/lib/Interfaces';
-import { IPluginLogger } from '@bettercorp/service-base/lib/interfaces/logger';
+import { Tools } from "@bettercorp/tools/lib/Tools";
+import { IServerConfig, webRequest as CWR } from "../../weblib";
+import moment = require("moment");
+import { IDictionary } from "@bettercorp/tools/lib/Interfaces";
+import { IPluginLogger } from "@bettercorp/service-base/lib/interfaces/logger";
 
 export enum IUCRMDirection {
-  Descending = 'DESC',
-  Ascending = 'ASC'
+  Descending = "DESC",
+  Ascending = "ASC",
 }
 export enum IUCRMServiceStatus {
   Prepared = 0,
@@ -24,7 +24,7 @@ export enum IUCRMInvoiceStatus {
   PartiallyPaid = 2,
   Paid = 3,
   Void = 4,
-  ProcessedProforma = 5
+  ProcessedProforma = 5,
 }
 
 export interface UCRM_v2_GetInvoices {
@@ -60,54 +60,105 @@ export class UCRM implements IUCRM {
   getServicePlanSurcharges(serviceId?: number, id?: number): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest(`/clients/services/${ serviceId }/service-surcharges${ Tools.isNullOrUndefined(id) ? '' : `/${ id }` }`, 'GET').then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      self
+        .webRequest(
+          `/clients/services/${serviceId}/service-surcharges${
+            Tools.isNullOrUndefined(id) ? "" : `/${id}`
+          }`,
+          "GET"
+        )
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
   getServicePlans(id?: number): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/service-plans${ Tools.isNullOrUndefined(id) ? '' : `/${ id }` }`, 'GET').then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(
+          `/service-plans${Tools.isNullOrUndefined(id) ? "" : `/${id}`}`,
+          "GET"
+        )
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
 
-  private webRequest(path: string, method: string, params: Object | undefined = undefined, data: Object | undefined = undefined, additionalProps: Object | undefined = undefined) {
-    return CWR(this.ServerConfig, '/crm/api/v1.0', path, method, params, data, additionalProps);
+  private webRequest(
+    path: string,
+    method: string,
+    params: Object | undefined = undefined,
+    data: Object | undefined = undefined,
+    additionalProps: Object | undefined = undefined
+  ) {
+    return CWR(
+      this.ServerConfig,
+      "/crm/api/v1.0",
+      path,
+      method,
+      params,
+      data,
+      additionalProps
+    );
   }
 
-  addNewServiceForClient(service: UCRM_Service, clientId: number): Promise<any> {
+  addNewServiceForClient(
+    service: UCRM_Service,
+    clientId: number
+  ): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest(`/clients/${ clientId }/services`, 'POST', undefined, service).then(resolve).catch(reject);
+      self
+        .webRequest(`/clients/${clientId}/services`, "POST", undefined, service)
+        .then(resolve)
+        .catch(reject);
     });
   }
   addNewClient(client: UCRM_Client): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest(`/clients`, 'POST', undefined, client).then(resolve).catch(reject);
+      self
+        .webRequest(`/clients`, "POST", undefined, client)
+        .then(resolve)
+        .catch(reject);
     });
   }
   getPayments(clientId?: number): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/payments${ clientId !== undefined ? `?clientId=${ clientId }&limit=10000` : '?limit=10000' }`, 'GET').then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(
+          `/payments${
+            clientId !== undefined
+              ? `?clientId=${clientId}&limit=10000`
+              : "?limit=10000"
+          }`,
+          "GET"
+        )
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
   getPaymentMethods(): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/payment-methods`, 'GET').then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(`/payment-methods`, "GET")
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
-  getInvoicePdf(invoiceId: number, clientId: number): Promise<any> {
-    const self = this;
+  async getInvoicePdf(invoiceId: number, clientId: number): Promise<any> {
+    throw "NEED WORK";
+    /*const self = this;
     return new Promise((resolve, reject) => {
       self.getInvoices(invoiceId, clientId).then((invoiceObj: any) => {
         if (`${ invoiceObj.clientId }` !== `${ clientId }`)
@@ -120,28 +171,63 @@ export class UCRM implements IUCRM {
           resolve(writer);
         }).catch(reject);
       }).catch(reject);
-    });
+    });*/
   }
-  getServices(serviceId?: number, clientId?: number, status?: number, offset: number = 0, limit: number = 10000): Promise<UCRM_Service | Array<UCRM_Service>> {
+  getServices(
+    serviceId?: number,
+    clientId?: number,
+    status?: number,
+    offset: number = 0,
+    limit: number = 10000
+  ): Promise<UCRM_Service | Array<UCRM_Service>> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest((clientId !== undefined && clientId !== null && (serviceId === undefined || serviceId === null) ?
-        `/clients/services?clientId=${ clientId }&offset=${ offset }&limit=${ limit }${ Tools.isNullOrUndefined(status) ? '' : `&statuses=${ status }` }` :
-        `/clients/services${ serviceId !== undefined && serviceId !== null ?
-          `/${ serviceId }?offset=${ offset }&limit=${ limit }` :
-          `?offset=${ offset }&limit=${ limit }` }${ Tools.isNullOrUndefined(status) ? '' : `&statuses=${ status }` }`), 'GET').then(x => resolve(x as UCRM_Service | Array<UCRM_Service>)).catch(reject);
+      self
+        .webRequest(
+          clientId !== undefined &&
+            clientId !== null &&
+            (serviceId === undefined || serviceId === null)
+            ? `/clients/services?clientId=${clientId}&offset=${offset}&limit=${limit}${
+                Tools.isNullOrUndefined(status) ? "" : `&statuses=${status}`
+              }`
+            : `/clients/services${
+                serviceId !== undefined && serviceId !== null
+                  ? `/${serviceId}?offset=${offset}&limit=${limit}`
+                  : `?offset=${offset}&limit=${limit}`
+              }${Tools.isNullOrUndefined(status) ? "" : `&statuses=${status}`}`,
+          "GET"
+        )
+        .then((x) => resolve(x as UCRM_Service | Array<UCRM_Service>))
+        .catch(reject);
     });
   }
   getServiceSurcharges(serviceId: number): Promise<any[]> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest(`/clients/services/${ serviceId }/service-surcharges`, 'GET').then(x => resolve(x as Array<any>)).catch(reject);
+      self
+        .webRequest(`/clients/services/${serviceId}/service-surcharges`, "GET")
+        .then((x) => resolve(x as Array<any>))
+        .catch(reject);
     });
   }
   getInvoices(invoiceId?: number, clientId?: number): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      self.webRequest((clientId !== undefined && clientId !== null && (invoiceId === undefined || invoiceId === null) ? `/invoices?clientId=${ clientId }&limit=10000` : `/invoices${ invoiceId !== undefined && invoiceId !== null ? `/${ invoiceId }?limit=10000` : '?limit=10000' }`), 'GET').then(x => resolve(x as Array<any>)).catch(reject);
+      self
+        .webRequest(
+          clientId !== undefined &&
+            clientId !== null &&
+            (invoiceId === undefined || invoiceId === null)
+            ? `/invoices?clientId=${clientId}&limit=10000`
+            : `/invoices${
+                invoiceId !== undefined && invoiceId !== null
+                  ? `/${invoiceId}?limit=10000`
+                  : "?limit=10000"
+              }`,
+          "GET"
+        )
+        .then((x) => resolve(x as Array<any>))
+        .catch(reject);
     });
   }
   v2_getInvoices(query: UCRM_v2_GetInvoices): Promise<any> {
@@ -150,152 +236,291 @@ export class UCRM implements IUCRM {
       let req = `/invoices?`;
       if (Tools.isNullOrUndefined(query.id)) {
         let reqQRY = [];
-        let simpleProps = ['clientId', 'createdDateFrom', 'createdDateTo', 'number', 'query', 'limit', 'offset', 'order', 'direction'];
+        let simpleProps = [
+          "clientId",
+          "createdDateFrom",
+          "createdDateTo",
+          "number",
+          "query",
+          "limit",
+          "offset",
+          "order",
+          "direction",
+        ];
         for (let prop of simpleProps) {
           if (!Tools.isNullOrUndefined((query as IDictionary<any>)[prop]))
-            reqQRY.push(`${ prop }=${ encodeURIComponent((query as IDictionary<any>)[prop]) }`);
+            reqQRY.push(
+              `${prop}=${encodeURIComponent((query as IDictionary<any>)[prop])}`
+            );
         }
         if (!Tools.isNullOrUndefined(query.overdue))
-          reqQRY.push(`overdue=${ query.overdue ? '1' : '0' }`);
+          reqQRY.push(`overdue=${query.overdue ? "1" : "0"}`);
         if (!Tools.isNullOrUndefined(query.proforma))
-          reqQRY.push(`proforma=${ query.proforma ? '1' : '0' }`);
+          reqQRY.push(`proforma=${query.proforma ? "1" : "0"}`);
         if (!Tools.isNullOrUndefined(query.statuses))
           for (let status of query.statuses!)
-            reqQRY.push(`statuses[]=${ status }`);
-        req += reqQRY.join('&');
+            reqQRY.push(`statuses[]=${status}`);
+        req += reqQRY.join("&");
       } else {
-        req = `/invoices/${ query.id }`;
+        req = `/invoices/${query.id}`;
       }
 
       self.log.debug(req);
-      self.webRequest(req, 'GET').then(x => resolve(x as Array<any>)).catch(reject);
+      self
+        .webRequest(req, "GET")
+        .then((x) => resolve(x as Array<any>))
+        .catch(reject);
     });
   }
-  getClient(id?: number, emailOrPhoneNumber?: string, offset: number = 0, limit: number = 1000): Promise<any> {
+  getClient(
+    id?: number,
+    emailOrPhoneNumber?: string,
+    offset: number = 0,
+    limit: number = 1000
+  ): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
       if (Tools.isNullOrUndefined(emailOrPhoneNumber))
-        return self.webRequest(`/clients${ id ? `/${ id }` : `?limit=${ limit }&offset=${ offset }` }`, 'GET').then(x => resolve(x as Array<any> | any)).catch(reject);
+        return self
+          .webRequest(
+            `/clients${id ? `/${id}` : `?limit=${limit}&offset=${offset}`}`,
+            "GET"
+          )
+          .then((x) => resolve(x as Array<any> | any))
+          .catch(reject);
 
-      let tempEmailOrPhone = `${ emailOrPhoneNumber }`.toLowerCase();
-      return self.webRequest(`/clients?limit=${ limit }&offset=${ offset }`, 'GET').then(async (x) => {
-        for (let clientObj of (x as any[])) {
-          let clientContacts = await self.webRequest(`/clients/${ clientObj.id }/contacts`, 'GET');
-          for (let contactObj of (clientContacts as any[])) {
-            if (contactObj.email !== undefined && contactObj.email !== null && `${ contactObj.email }`.toLowerCase() === tempEmailOrPhone) {
-              return resolve(clientObj);
-            }
-            if (contactObj.phone !== undefined && contactObj.phone !== null && `${ contactObj.phone }`.toLowerCase() === tempEmailOrPhone) {
-              return resolve(clientObj);
+      let tempEmailOrPhone = `${emailOrPhoneNumber}`.toLowerCase();
+      return self
+        .webRequest(`/clients?limit=${limit}&offset=${offset}`, "GET")
+        .then(async (x) => {
+          for (let clientObj of x as any[]) {
+            let clientContacts = await self.webRequest(
+              `/clients/${clientObj.id}/contacts`,
+              "GET"
+            );
+            for (let contactObj of clientContacts as any[]) {
+              if (
+                contactObj.email !== undefined &&
+                contactObj.email !== null &&
+                `${contactObj.email}`.toLowerCase() === tempEmailOrPhone
+              ) {
+                return resolve(clientObj);
+              }
+              if (
+                contactObj.phone !== undefined &&
+                contactObj.phone !== null &&
+                `${contactObj.phone}`.toLowerCase() === tempEmailOrPhone
+              ) {
+                return resolve(clientObj);
+              }
             }
           }
-        }
-        resolve(null);
-      }).catch(reject);
+          resolve(null);
+        })
+        .catch(reject);
     });
   }
   setClient(id: number, clientObj: any): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/clients/${ id }`, 'PATCH', undefined, clientObj).then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(`/clients/${id}`, "PATCH", undefined, clientObj)
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
   setService(id: number, serviceObj: any): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/clients/services/${ id }`, 'PATCH', undefined, serviceObj).then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(`/clients/services/${id}`, "PATCH", undefined, serviceObj)
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
-  addPayment(clientId: number, methodId: string, amount: number, note: string, invoiceIds?: number[], applyToInvoicesAutomatically?: boolean, userId?: number, additionalProps?: any): Promise<any> {
+  addPayment(
+    clientId: number,
+    methodId: string,
+    amount: number,
+    note: string,
+    invoiceIds?: number[],
+    applyToInvoicesAutomatically?: boolean,
+    userId?: number,
+    additionalProps?: any
+  ): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
       let nowFormat = moment().format();
       let sendObj: any = {
         clientId,
         amount,
-        currencyCode: 'ZAR',
+        currencyCode: "ZAR",
         userId,
         note,
         methodId,
         invoiceIds,
         createdDate: nowFormat,
         providerPaymentTime: nowFormat,
-        ...(additionalProps || {})
+        ...(additionalProps || {}),
       };
-      return self.webRequest(`/payments`, 'POST', undefined, sendObj).then(async (x) => {
-        resolve(x);
-      }).catch(x => reject(x.response.data || x));
+      return self
+        .webRequest(`/payments`, "POST", undefined, sendObj)
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch((x) => reject(x.response.data || x));
     });
   }
   getClientBankAccount(id?: number, clientId?: number): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(id !== undefined ? `/clients/bank-accounts/${ id }` : `/clients/${ clientId }/bank-accounts`, 'GET').then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(
+          id !== undefined
+            ? `/clients/bank-accounts/${id}`
+            : `/clients/${clientId}/bank-accounts`,
+          "GET"
+        )
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
   addClientBankAccount(clientId: number, obj: any): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/clients/${ clientId }/bank-accounts`, 'POST', undefined, obj).then(async (x) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(
+          `/clients/${clientId}/bank-accounts`,
+          "POST",
+          undefined,
+          obj
+        )
+        .then(async (x) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
-  addNewInvoice(items: Array<UCRM_InvoiceItem>, attributes: Array<UCRM_InvoiceAttribute>, maturityDays: number = 14,
-    invoiceTemplateId: number, clientId: number, applyCredit: Boolean = true,
-    proforma: boolean = false, adminNotes?: string, notes?: string): Promise<any> {
+  addNewInvoice(
+    items: Array<UCRM_InvoiceItem>,
+    attributes: Array<UCRM_InvoiceAttribute>,
+    maturityDays: number = 14,
+    invoiceTemplateId: number,
+    clientId: number,
+    applyCredit: Boolean = true,
+    proforma: boolean = false,
+    adminNotes?: string,
+    notes?: string
+  ): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/clients/${ clientId }/invoices`, 'POST', undefined, {
-        items,
-        attributes,
-        maturityDays,
-        invoiceTemplateId,
-        applyCredit,
-        proforma,
-        adminNotes,
-        notes
-      }).then(async (x: any) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(`/clients/${clientId}/invoices`, "POST", undefined, {
+          items,
+          attributes,
+          maturityDays,
+          invoiceTemplateId,
+          applyCredit,
+          proforma,
+          adminNotes,
+          notes,
+        })
+        .then(async (x: any) => {
+          resolve(x);
+        })
+        .catch(reject);
     });
   }
   sendInvoice(invoiceId: string): Promise<any> {
     let self = this;
     return new Promise((resolve, reject) => {
-      return self.webRequest(`/invoices/${ invoiceId }/send`, 'PATCH').then(async (x: any) => {
-        resolve(x);
-      }).catch(reject);
+      return self
+        .webRequest(`/invoices/${invoiceId}/send`, "PATCH")
+        .then(async (x: any) => {
+          resolve(x);
+        })
+        .catch(reject);
+    });
+  }
+  getCountries(): Promise<Array<UCRM_Country>> {
+    let self = this;
+    return new Promise((resolve, reject) => {
+      return self
+        .webRequest(`/countries`, "GET")
+        .then(async (x) => {
+          resolve(x as any as Array<UCRM_Country>);
+        })
+        .catch(reject);
     });
   }
 }
 
 export interface IUCRM {
-  addNewInvoice(items: Array<UCRM_InvoiceItem>, attributes: Array<UCRM_InvoiceAttribute>, maturityDays: number, invoiceTemplateId: number, clientId: number, applyCredit?: Boolean, proforma?: boolean, adminNotes?: string, notes?: string): Promise<any>;
+  addNewInvoice(
+    items: Array<UCRM_InvoiceItem>,
+    attributes: Array<UCRM_InvoiceAttribute>,
+    maturityDays: number,
+    invoiceTemplateId: number,
+    clientId: number,
+    applyCredit?: Boolean,
+    proforma?: boolean,
+    adminNotes?: string,
+    notes?: string
+  ): Promise<any>;
   sendInvoice(invoiceId: string): Promise<any>;
   addNewServiceForClient(service: UCRM_Service, clientId: number): Promise<any>;
   addNewClient(client: UCRM_Client): Promise<any>;
   getPayments(clientId?: number): Promise<Array<any> | any>;
   getPaymentMethods(): Promise<Array<any> | any>;
   getInvoicePdf(invoiceId: number, clientId: number): Promise<any>;
-  getServices(serviceId?: number, clientId?: number, status?: number, offset?: number, limit?: number): Promise<UCRM_Service | Array<UCRM_Service>>;
+  getServices(
+    serviceId?: number,
+    clientId?: number,
+    status?: number,
+    offset?: number,
+    limit?: number
+  ): Promise<UCRM_Service | Array<UCRM_Service>>;
   getServiceSurcharges(serviceId: number): Promise<Array<any>>;
   getInvoices(invoiceId?: number, clientId?: number): Promise<Array<any> | any>;
-  getClient(id?: number, emailOrPhoneNumber?: string, offset?: number, limit?: number): Promise<Array<any> | any>;
+  getClient(
+    id?: number,
+    emailOrPhoneNumber?: string,
+    offset?: number,
+    limit?: number
+  ): Promise<Array<any> | any>;
   setClient(id: number, clientObj: any): Promise<Array<any> | any>;
-  addPayment(clientId: number, methodId: string, amount: number, note: string, invoiceIds?: Array<number>, applyToInvoicesAutomatically?: boolean, userId?: number, additionalProps?: any): Promise<Array<any> | any>;
-  getClientBankAccount(id?: number, clientId?: number): Promise<Array<any> | any>;
+  addPayment(
+    clientId: number,
+    methodId: string,
+    amount: number,
+    note: string,
+    invoiceIds?: Array<number>,
+    applyToInvoicesAutomatically?: boolean,
+    userId?: number,
+    additionalProps?: any
+  ): Promise<Array<any> | any>;
+  getClientBankAccount(
+    id?: number,
+    clientId?: number
+  ): Promise<Array<any> | any>;
   addClientBankAccount(clientId: number, obj: any): Promise<Array<any> | any>;
   getServicePlans(id?: number): Promise<Array<any> | any>;
-  getServicePlanSurcharges(serviceId?: number, id?: number): Promise<Array<any> | any>;
-  getServicePlanSurcharges(serviceId?: number, id?: number): Promise<Array<any> | any>;
+  getServicePlanSurcharges(
+    serviceId?: number,
+    id?: number
+  ): Promise<Array<any> | any>;
+  getServicePlanSurcharges(
+    serviceId?: number,
+    id?: number
+  ): Promise<Array<any> | any>;
   setService(id: number, serviceObj: any): Promise<any>;
+  getCountries(): Promise<Array<UCRM_Country>>;
   //getTickets (id?: number): Promise<Array<any> | any>;
   //setTicket (ticketData: any, id?: number): Promise<any>;
   //getJobs (id?: number): Promise<Array<any> | any>;
@@ -303,21 +528,21 @@ export interface IUCRM {
 }
 
 export enum UCRM_ServicePlanType {
-  Internet = 'Internet',
-  General = 'General'
+  Internet = "Internet",
+  General = "General",
 }
 export enum UCRM_DiscountType {
   NoDiscount = 0,
   Percentage = 1,
-  Fixed = 2
+  Fixed = 2,
 }
 export enum UCRM_ContractLengthType {
   OpenEnd = 1,
-  CloseEnd = 2
+  CloseEnd = 2,
 }
 export enum UCRM_InvoicingPeriodType {
   Backwards = 1,
-  Forwards = 2
+  Forwards = 2,
 }
 export enum UCRM_ServiceStatus {
   Prepared = 0,
@@ -383,15 +608,17 @@ export interface UCRM_Service {
   hasOutage?: boolean;
   unmsClientSiteId?: string;
   lastInvoicedDate?: string;
-  attributes: [{
-    value: string;
-    customAttributeId: number;
-    id: string;
-    serviceId: number;
-    name: string;
-    key: string;
-    clientZoneVisible?: boolean;
-  }];
+  attributes: [
+    {
+      value: string;
+      customAttributeId: number;
+      id: string;
+      serviceId: number;
+      name: string;
+      key: string;
+      clientZoneVisible?: boolean;
+    }
+  ];
   suspensionReasonId?: number;
   serviceChangeRequestId?: string;
   setupFeePrice?: number;
@@ -443,7 +670,7 @@ export enum UCRM_Service_DiscountType {
 }
 export enum UCRM_Client_Type {
   Residential = 1,
-  Company = 2
+  Company = 2,
 }
 export interface UCRM_InvoiceItem {
   label: string;
@@ -458,4 +685,9 @@ export interface UCRM_InvoiceItem {
 export interface UCRM_InvoiceAttribute {
   value: string;
   customAttributeId: number | null;
+}
+export interface UCRM_Country {
+  id: number;
+  name: string;
+  code: string;
 }
